% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfUtils.R
\name{clusterProfilerWriteGO}
\alias{clusterProfilerWriteGO}
\title{Write clusterProfile GO data to csv}
\usage{
clusterProfilerWriteGO(ego, csvout, converter = EntrezToSymbolHuman)
}
\arguments{
\item{csvout}{Path to output csv,}

\item{converter}{Function to map ENtrez IDs to Gene Symbol (Default=EntrezToSymbolHuman)}

\item{results}{ekg clusterProfiler enrichGO object.}
}
\description{
Write clusterProfile GO data to csv
}
