% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goUtils.R
\name{clusterProfilerWriteKEGG}
\alias{clusterProfilerWriteKEGG}
\title{Write clusterProfile KEGG data to csv}
\usage{
clusterProfilerWriteKEGG(ekg, csvout, converter = EntrezToSymbolHuman)
}
\arguments{
\item{csvout}{Path to output csv,}

\item{converter}{Function to map ENtrez IDs to Gene Symbol (Default=EntrezToSymbolHuman)}

\item{results}{ekg clusterProfiler enrichKEGG object.}
}
\description{
Write clusterProfile KEGG data to csv
}
