% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{DESeqPlotHeatmap}
\alias{DESeqPlotHeatmap}
\title{Draw sample to sample heatmap using DESeq2 object.}
\usage{
DESeqPlotHeatmap(dds, col_df, col_labels, title, filename)
}
\arguments{
\item{col_df}{Design dataframe}

\item{col_labels}{Label names for columns}

\item{title}{Plot title}

\item{filename}{Path to file}

\item{dds.vst}{DESeq2 post VST obj .}
}
\description{
Draw sample to sample heatmap using DESeq2 object.
}
