% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{DESeqPlotPCA}
\alias{DESeqPlotPCA}
\title{Draw PCA using DESeq2 object.}
\usage{
DESeqPlotPCA(dds.vst, intgroup, title)
}
\arguments{
\item{dds.vst}{DESeq2 post VST obj .}

\item{intrgroup}{Features to plot (intgroup[1] - color, intgroup[2]-shape)}
}
\description{
Draw PCA using DESeq2 object.
}
