% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{DESeqPlotSampleDistmap}
\alias{DESeqPlotSampleDistmap}
\title{Draw sample to sample heatmap using DESeq2 object.}
\usage{
DESeqPlotSampleDistmap(dds.vst, filename = NULL)
}
\arguments{
\item{dds.vst}{DESeq2 post VST obj .}

\item{filename}{Filename to save to}
}
\description{
Draw sample to sample heatmap using DESeq2 object.
}
