% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{DESeqPlotVolcano}
\alias{DESeqPlotVolcano}
\title{Draw volcano plot using DESeq2 results object.}
\usage{
DESeqPlotVolcano(results, numerator, denominator, pValCutoff = 1)
}
\arguments{
\item{results}{DESeq2 results obj.}

\item{numerator}{Name of numerator  of contrast}

\item{denominator}{Name of denominator  of contrast}

\item{pValCutoff}{Cutoff to draw the lines}
}
\value{
plotObject
}
\description{
Draw volcano plot using DESeq2 results object.
}
