% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goUtils.R
\name{EntrezToSymbolHuman}
\alias{EntrezToSymbolHuman}
\title{Convert Entrez ID to Gene symbol (Human)}
\usage{
EntrezToSymbolHuman(entrezID)
}
\arguments{
\item{entrezID}{Entrez ID (Human),}
}
\value{
Gene Symbol
}
\description{
Convert Entrez ID to Gene symbol (Human)
}
