% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboraptorUtils.R
\name{extendLength}
\alias{extendLength}
\title{Extend any vector to specified length filled with NAs}
\usage{
extendLength(vector, newLength)
}
\arguments{
\item{vector}{Input vector.}

\item{newLength}{Final length of extended vecotr}
}
\value{
Extended vector
}
\description{
Extend any vector to specified length filled with NAs
}
