% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboraptorUtils.R
\name{riboSummaryCollapseDF}
\alias{riboSummaryCollapseDF}
\title{Extend any vector to specified length filled with NAs}
\usage{
riboSummaryCollapseDF(df)
}
\arguments{
\item{df}{Data frame as read from riboSummary Counts file}
}
\value{
Counts Matrix
}
\description{
Extend any vector to specified length filled with NAs
}
