% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboraptorUtils.R
\name{riboSummaryParseCountColumn}
\alias{riboSummaryParseCountColumn}
\title{Parse count column of dataframe saved by riboraptor}
\usage{
riboSummaryParseCountColumn(col, newLength)
}
\arguments{
\item{col}{Count column,}

\item{newLength}{Final length of extended vecotr}
}
\value{
Extended vector post columns parsing
}
\description{
Parse count column of dataframe saved by riboraptor
}
